Put graph algos in separate file
